🔨 : 생각해보니 HashMap보다 Array리스트 쓰는게 훨씬 좋은거같아서 갈아엎음

+++ src/commerce/Customer.java
@@ -4,16 +4,16 @@ public class Customer {
     String name;
     String email;
     Rating rating;
-    double totalOrderAmount;
+    int totalOrderAmount;
 
-    public Customer(String name, String email, double totalOrderAmount) {
+    public Customer(String name, String email, int totalOrderAmount) {
         this.name = name;
         this.email = email;
         this.totalOrderAmount = totalOrderAmount;
         this.rating = rating(totalOrderAmount);
     }
 
-    Rating rating(double totalOrderAmount){
+    Rating rating(int totalOrderAmount){
         if (totalOrderAmount < 500_000) return Rating.BRONZE;
         if (totalOrderAmount < 1_000_000) return  Rating.SILVER;
         if (totalOrderAmount < 1_500_000) return  Rating.GOLD;
